Fix top-level error overlay rendering to prevent hydration errors
